# lib/api-spec/openapi.yaml
openapi: "3.0.0"
info:
  title: ACE-2026 API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.ace-2026.com
    description: Production Server
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
      required:
        - success
        - data
        - error
        - code
        - timestamp
    MultiLingual:
      type: object
      properties:
        en: { type: string }
        es: { type: string }
        fr: { type: string }
        zh: { type: string }
        ja: { type: string }
        ko: { type: string }
      required:
        - en
        - es
        - fr
        - zh
        - ja
        - ko
    MediaAsset:
      type: object
      properties:
        url: { type: string, format: uri }
        blurHash: { type: string }
        width: { type: number }
        height: { type: number }
        format: { type: string, enum: [webp, jpg, png] }
        dominantColors: { type: array, items: { type: string } }
        vibrantPalette: { $ref: "#/components/schemas/VibrantPalette", nullable: true }
      required:
        - url
        - blurHash
        - width
        - height
        - format
        - dominantColors
    VibrantPalette:
      type: object
      properties:
        vibrant: { type: string }
        muted: { type: string }
        darkVibrant: { type: string }
        darkMuted: { type: string }
        lightVibrant: { type: string }
        lightMuted: { type: string }
      required:
        - vibrant
        - muted
        - darkVibrant
        - darkMuted
        - lightVibrant
        - lightMuted
    ComposerIdentity:
      type: object
      properties:
        id: { type: string, format: uuid, nullable: true }
        name: { $ref: "#/components/schemas/MultiLingual", nullable: true }
        tagline: { $ref: "#/components/schemas/MultiLingual", nullable: true }
        biography: { $ref: "#/components/schemas/MultiLingual", nullable: true }
        awards: { type: array, items: { $ref: "#/components/schemas/MultiLingual" }, nullable: true }
        studioAddress: { $ref: "#/components/schemas/MultiLingual", nullable: true }
        portrait: { $ref: "#/components/schemas/MediaAsset", nullable: true }
        logo: { $ref: "#/components/schemas/MediaAsset", nullable: true }
        heroVideo: { type: string, nullable: true }
        socialLinks: { $ref: "#/components/schemas/SocialLinks", nullable: true }
        projects: { type: array, items: { $ref: "#/components/schemas/Project" }, nullable: true }
    SocialLinks:
      type: object
      properties:
        spotify: { type: string, format: uri, nullable: true }
        imdb: { type: string, format: uri, nullable: true }
        instagram: { type: string, format: uri, nullable: true }
        youtube: { type: string, format: uri, nullable: true }
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { $ref: "#/components/schemas/MultiLingual" }
        type: { type: string, enum: [film, game, animation, documentary] }
        year: { type: number }
        description: { $ref: "#/components/schemas/MultiLingual" }
        coverImage: { $ref: "#/components/schemas/MediaAsset", nullable: true }
      required:
        - id
        - title
        - type
        - year
        - description
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/auth/login:
    post:
      summary: Authenticate user and issue JWTs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required:
                - username
                - password
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/auth/logout:
    post:
      summary: Logout user and clear JWT cookies
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successful logout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/auth/refresh:
    post:
      summary: Refresh access token using refresh token
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Access token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/identity:
    get:
      summary: Get composer identity
      responses:
        "200":
          description: Composer identity data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    put:
      summary: Update composer identity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComposerIdentity"
      responses:
        "200":
          description: Composer identity updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /api/identity/translate:
    post:
      summary: Translate multilingual content using AI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text: { type: string }
                sourceLang: { $ref: "#/components/schemas/Locale" }
                fieldType: { type: string }
              required:
                - text
                - sourceLang
                - fieldType
      responses:
        "200":
          description: Translated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...